7.c

+++ 7.c
@@ -5,11 +5,9 @@ int main() {
     int *ptr1 = &a;
     int *ptr2 = &b;
 
-    
     printf("Integer division: %d\n", *ptr1 / *ptr2);
-
-    
-    printf("Floating-point division: %f\n", (double)(*ptr1) / *ptr2);
+    printf("Floating-point division: %.2f\n", (float)(*ptr1) / *ptr2);
 
     return 0;
 }
+
